post

+++ app/Http/Controllers/TasksController.php
@@ -53,10 +53,10 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
 
-        $task = new Task;
-        $task->status = $request->status;    
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
         return redirect('/');
     }
@@ -98,17 +98,19 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request)
     {
         $this->validate($request, [
             'status' => 'required|max:10',   
             'content' => 'required|max:255',
         ]);
 
-        $task = Task::find($id);
-        $task->status = $request->status;    
-        $task->content = $request->content;
-        $task->save();
+         if (\Auth::user()->id === $task->user_id) {
+        $request->user()->tasks()->update([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+        }
 
         return redirect('/');
     }
@@ -121,9 +123,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-
-        return redirect('/');
+        $task = App\Task::find($id);
+        
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
+        
+        return redirect()->back();
     }
 }
